network: Log errors from network failures
Before, we didn't display the errors that caused us to infer that the
network was broken. However, the errors are useful for debugging network
test failures -- for example, an error might arise because of problems
with the `quilt ssh` command, and not the network itself.

This patch also mildly simplifies the testing logic by not
separating failures to execute the DNS resolution command, and resolving
DNS entries to the wrong IP.