quilt-tester: Remove references to provider specific functions
Before, quilt-tester attempted to protect against leaked instances by
using provider-specific functions to ensure machines are stopped. This
does not offer particularly good protection because the Quilt cluster
uses the same API calls.

A better solution would be to have a completely separate script that
checks for leaked instances. For now, we remove the provider-specific
functions for simplicity, as it allows us to easily extend quilt-tester
to support providers other than Amazon.