duplicate-cluster: Don't double count reconnection attempts
Before, we simply counted the number of lines containing the string
"Registering worker" to determine the number of Spark workers
connecting to a master. However, sometimes when a worker attempts to
reconnect, it produces repeated lines in the log. We now parse the
master logs, and only count unique worker IP addresses.