quilt-tester: Add flag for skipping tests
Some tests should not be run against certain Quilt versions because the
tested features are not part of the release. For example, the recent
load balancing test should only be run against the dev version of Quilt,
and not version 0.1.0.

This commit adds the `-skip` flag for specifying tests to skip. It takes
the test suite name, and can be repeated:
`./tester -skip 10-network -skip load-balancer