tests: Move integration tests into main repository
This commit copies the integration tests defined in `quilt/tester` into
the main Quilt source repo. This way, integration tests will be more
tightly linked with the code it's testing. For example, new features can
be merged in the same PR as the accompanying integration test, and
integration tests will be versioned with releases.