dockerfile: Build off of official LTS version
The `jenkinsci/jenkins` image has been deprecated in favor of
`jenkins/jenkins` (which is managed by the official Jenkins community).

This commit also changes the Jenkins version to the LTS version, which
is useful because each LTS version has its own update center for
plugins. Before, every time we reinstalled the plugins (i.e. we rebuilt
the Docker image), the latest versions of the plugins were installed
without consideration for whether they were compatible with the Jenkins
version.