jenkins: Add option for disabling project after a test fails
This commit adds a boolean to the integration-tester jobs that allows
all future jobs to be disabled if a run fails. This way, broken clusters
won't get torn down if there's a job scheduled to run after (the
scheduled job won't get to run since the entire project is disabled).